F/Introduction-to-gorm: Practice Gorm fundamentals - Saved local domain object - Output local domain object that was saved - Queried and displayed specific domain objects that were saved
