+++ grails-app/domain/grailspractice/Person.groovy
@@ -1,0 +1,8 @@
+package grailspractice
+
+class Person {
+
+    String firstName
+    String lastName
+    int age
+}
